mcp-server/pyproject.toml
[project]
name = "npl-mcp"
version = "0.1.0"
description = "NPL MCP Server with artifact management and persona chat system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/npl_mcp"]

[project.scripts]
npl-mcp = "npl_mcp.server:main"
